Server Configuration File
Sample Server Configuration File needed to connect to the server. The
file has to be placed on the server with an .ovpn extension so that the
OpenVPN tool can understand it. This holds all the details of the
location of the certificates and keys required for a secured
communication between client and server.